+++ packages/expo-document-picker/plugin/src/index.ts
@@ -1,1 +1,3 @@
-export default () => 'expo-document-picker';
+import { IosProps } from './withDocumentPickerIOS';
+
+export default (props: IosProps = {}): [string, IosProps] => ['expo-document-picker', props];